+++ packages/moonstone/VirtualList/VirtualList.js
@@ -86,94 +86,95 @@ let VirtualList = ({itemSize, role, ...rest}) => {
 
 VirtualList.displayName = 'VirtualList';
 
+// TBD: indentation is broken intentionally to help comparing
 VirtualList.propTypes = /** @lends moonstone/VirtualList.VirtualList.prototype */ {
-	/**
-	 * Size of an item for the VirtualList; valid value is a number generally.
-	 * For different item size, value is an object that has `minSize`
-	 * and `size` as properties.
-	 * If the direction for the list is vertical, itemSize means the height of an item.
-	 * For horizontal, it means the width of an item.
-	 *
-	 * Usage:
-	 * ```
-	 * <VirtualList itemSize={ri.scale(72)} />
-	 * ```
-	 *
-	 * @type {Number|ui/VirtualList.itemSizesShape}
-	 * @required
-	 * @public
-	 */
-	itemSize: PropTypes.oneOfType([PropTypes.number, itemSizesShape]).isRequired,
+		/**
+		 * Size of an item for the VirtualList; valid value is a number generally.
+		 * For different item size, value is an object that has `minSize`
+		 * and `size` as properties.
+		 * If the direction for the list is vertical, itemSize means the height of an item.
+		 * For horizontal, it means the width of an item.
+		 *
+		 * Usage:
+		 * ```
+		 * <VirtualList itemSize={ri.scale(72)} />
+		 * ```
+		 *
+		 * @type {Number|ui/VirtualList.itemSizesShape}
+		 * @required
+		 * @public
+		 */
+		itemSize: PropTypes.oneOfType([PropTypes.number, itemSizesShape]).isRequired,
 
-	cbScrollTo: PropTypes.func,
+		cbScrollTo: PropTypes.func,
 
-	/**
-	 * `false` if the content of the list or the scroller could get focus
-	 *
-	 * @type {Boolean}
-	 * @default false
-	 * @private
-	 */
-	'data-spotlight-container-disabled': PropTypes.bool,
+		/**
+		 * `false` if the content of the list or the scroller could get focus
+		 *
+		 * @type {Boolean}
+		 * @default false
+		 * @private
+		 */
+		'data-spotlight-container-disabled': PropTypes.bool,
 
-	direction: PropTypes.oneOf(['horizontal', 'vertical']),
-	focusableScrollbar: PropTypes.bool,
+		direction: PropTypes.oneOf(['horizontal', 'vertical']),
+		focusableScrollbar: PropTypes.bool,
 
-	/**
-	 * Specifies how to show horizontal scrollbar.
-	 *
-	 * Valid values are:
-	 * * `'auto'`,
-	 * * `'visible'`, and
-	 * * `'hidden'`.
-	 *
-	 * @type {String}
-	 * @default 'auto'
-	 * @public
-	 */
-	horizontalScrollbar: PropTypes.oneOf(['auto', 'visible', 'hidden']),
+		/**
+		 * Specifies how to show horizontal scrollbar.
+		 *
+		 * Valid values are:
+		 * * `'auto'`,
+		 * * `'visible'`, and
+		 * * `'hidden'`.
+		 *
+		 * @type {String}
+		 * @default 'auto'
+		 * @public
+		 */
+		horizontalScrollbar: PropTypes.oneOf(['auto', 'visible', 'hidden']),
 
-	itemSizes: PropTypes.array,
+		itemSizes: PropTypes.array,
 
-	/**
-	 * Specifies overscroll effects shows on which type of inputs.
-	 *
-	 * @type {Object}
-	 * @default {
-	 *	arrowKey: false,
-	 *	drag: false,
-	 *	pageKey: false,
-	 *	scrollbarButton: false,
-	 *	wheel: true
-	 * }
-	 * @private
-	 */
-	overscrollEffectOn: PropTypes.shape({
-		arrowKey: PropTypes.bool,
-		drag: PropTypes.bool,
-		pageKey: PropTypes.bool,
-		scrollbarButton: PropTypes.bool,
-		wheel: PropTypes.bool
-	}),
+		/**
+		 * Specifies overscroll effects shows on which type of inputs.
+		 *
+		 * @type {Object}
+		 * @default {
+		 *	arrowKey: false,
+		*	drag: false,
+		*	pageKey: false,
+		*	scrollbarButton: false,
+		*	wheel: true
+		* }
+		* @private
+		*/
+		overscrollEffectOn: PropTypes.shape({
+			arrowKey: PropTypes.bool,
+			drag: PropTypes.bool,
+			pageKey: PropTypes.bool,
+			scrollbarButton: PropTypes.bool,
+			wheel: PropTypes.bool
+		}),
 
-	preventBubblingOnKeyDown: PropTypes.oneOf(['none', 'programmatic']),
-	role: PropTypes.string,
+		preventBubblingOnKeyDown: PropTypes.oneOf(['none', 'programmatic']),
+		role: PropTypes.string,
 
-	type: PropTypes.string,
+		type: PropTypes.string,
 
-	/**
-	 * Specifies how to show vertical scrollbar.
-	 *
-	 * Valid values are:
-	 * * `'auto'`,
-	 * * `'visible'`, and
-	 * * `'hidden'`.
-	 *
-	 * @type {String}
-	 * @default 'auto'
-	 * @public
-	 */
-	verticalScrollbar: PropTypes.oneOf(['auto', 'visible', 'hidden'])
+		/**
+		 * Specifies how to show vertical scrollbar.
+		 *
+		 * Valid values are:
+		 * * `'auto'`,
+		 * * `'visible'`, and
+		 * * `'hidden'`.
+		 *
+		 * @type {String}
+		 * @default 'auto'
+		 * @public
+		 */
+		verticalScrollbar: PropTypes.oneOf(['auto', 'visible', 'hidden'])
 };
 
 VirtualList.defaultProps = {
@@ -257,94 +258,95 @@ let VirtualGridList = ({role, ...rest}) => {
 
 VirtualGridList.displayName = 'VirtualGridList';
 
+// TBD: indentation is broken intentionally to help comparing
 VirtualGridList.propTypes = /** @lends moonstone/VirtualList.VirtualGridList.prototype */ {
-	/**
-	 * Size of an item for the VirtualGridList; valid value is an object that has `minWidth`
-	 * and `minHeight` as properties.
-	 *
-	 * Usage:
-	 * ```
-	 * <VirtualGridList
-	 * 	itemSize={{
-	 * 		minWidth: ri.scale(180),
-	 * 		minHeight: ri.scale(270)
-	 * 	}}
-	 * />
-	 * ```
-	 *
-	 * @type {ui/VirtualList.gridListItemSizeShape}
-	 * @required
-	 * @public
-	 */
-	itemSize: gridListItemSizeShape.isRequired,
+		/**
+		 * Size of an item for the VirtualGridList; valid value is an object that has `minWidth`
+		 * and `minHeight` as properties.
+		 *
+		 * Usage:
+		 * ```
+		 * <VirtualGridList
+		 * 	itemSize={{
+		 * 		minWidth: ri.scale(180),
+		 * 		minHeight: ri.scale(270)
+		 * 	}}
+		 * />
+		 * ```
+		 *
+		 * @type {ui/VirtualList.gridListItemSizeShape}
+		 * @required
+		 * @public
+		 */
+		itemSize: gridListItemSizeShape.isRequired,
 
-	cbScrollTo: PropTypes.func,
+		cbScrollTo: PropTypes.func,
 
-	/**
-	 * `false` if the content of the list or the scroller could get focus
-	 *
-	 * @type {Boolean}
-	 * @default false
-	 * @private
-	 */
-	'data-spotlight-container-disabled': PropTypes.bool,
+		/**
+		 * `false` if the content of the list or the scroller could get focus
+		 *
+		 * @type {Boolean}
+		 * @default false
+		 * @private
+		 */
+		'data-spotlight-container-disabled': PropTypes.bool,
 
-	direction: PropTypes.oneOf(['horizontal', 'vertical']),
-	focusableScrollbar: PropTypes.bool,
-	/**
-	 * Specifies how to show horizontal scrollbar.
-	 *
-	 * Valid values are:
-	 * * `'auto'`,
-	 * * `'visible'`, and
-	 * * `'hidden'`.
-	 *
-	 * @type {String}
-	 * @default 'auto'
-	 * @public
-	 */
-	horizontalScrollbar: PropTypes.oneOf(['auto', 'visible', 'hidden']),
+		direction: PropTypes.oneOf(['horizontal', 'vertical']),
+		focusableScrollbar: PropTypes.bool,
+		/**
+		 * Specifies how to show horizontal scrollbar.
+		 *
+		 * Valid values are:
+		 * * `'auto'`,
+		 * * `'visible'`, and
+		 * * `'hidden'`.
+		 *
+		 * @type {String}
+		 * @default 'auto'
+		 * @public
+		 */
+		horizontalScrollbar: PropTypes.oneOf(['auto', 'visible', 'hidden']),
 
-	itemSizes: PropTypes.array,
+		itemSizes: PropTypes.array,
 
-	/**
-	 * Specifies overscroll effects shows on which type of inputs.
-	 *
-	 * @type {Object}
-	 * @default {
-	 *	arrowKey: false,
-	 *	drag: false,
-	 *	pageKey: false,
-	 *	scrollbarButton: false,
-	 *	wheel: true
-	 * }
-	 * @private
-	 */
-	overscrollEffectOn: PropTypes.shape({
-		arrowKey: PropTypes.bool,
-		drag: PropTypes.bool,
-		pageKey: PropTypes.bool,
-		scrollbarButton: PropTypes.bool,
-		wheel: PropTypes.bool
-	}),
+		/**
+		 * Specifies overscroll effects shows on which type of inputs.
+		 *
+		 * @type {Object}
+		 * @default {
+		 *	arrowKey: false,
+		*	drag: false,
+		*	pageKey: false,
+		*	scrollbarButton: false,
+		*	wheel: true
+		* }
+		* @private
+		*/
+		overscrollEffectOn: PropTypes.shape({
+			arrowKey: PropTypes.bool,
+			drag: PropTypes.bool,
+			pageKey: PropTypes.bool,
+			scrollbarButton: PropTypes.bool,
+			wheel: PropTypes.bool
+		}),
 
-	preventBubblingOnKeyDown: PropTypes.oneOf(['none', 'programmatic']),
-	role: PropTypes.string,
-	type: PropTypes.string,
+		preventBubblingOnKeyDown: PropTypes.oneOf(['none', 'programmatic']),
+		role: PropTypes.string,
+		type: PropTypes.string,
 
-	/**
-	 * Specifies how to show vertical scrollbar.
-	 *
-	 * Valid values are:
-	 * * `'auto'`,
-	 * * `'visible'`, and
-	 * * `'hidden'`.
-	 *
-	 * @type {String}
-	 * @default 'auto'
-	 * @public
-	 */
-	verticalScrollbar: PropTypes.oneOf(['auto', 'visible', 'hidden'])
+		/**
+		 * Specifies how to show vertical scrollbar.
+		 *
+		 * Valid values are:
+		 * * `'auto'`,
+		 * * `'visible'`, and
+		 * * `'hidden'`.
+		 *
+		 * @type {String}
+		 * @default 'auto'
+		 * @public
+		 */
+		verticalScrollbar: PropTypes.oneOf(['auto', 'visible', 'hidden'])
 };
 
 VirtualGridList.defaultProps = {
