--- packages/moonstone/VirtualList/virtualListItemsRenderer.js
@@ -1,38 +1,0 @@
-import React from 'react';
-
-/* eslint-disable enact/prop-types */
-const listItemsRenderer = (props) => {
-	const {
-		cc,
-		handlePlaceholderFocus,
-		itemContainerRef: initUiItemContainerRef,
-		needsScrollingPlaceholder,
-		primary,
-		role,
-		SpotlightPlaceholder
-	} = props;
-
-	return (
-		<>
-			{cc.length ? (
-				<div ref={initUiItemContainerRef} role={role}>{cc}</div>
-			) : null}
-			{primary ? null : (
-				<SpotlightPlaceholder
-					data-index={0}
-					data-vl-placeholder
-					// a zero width/height element can't be focused by spotlight so we're giving
-					// the placeholder a small size to ensure it is navigable
-					onFocus={handlePlaceholderFocus}
-					style={{width: 10}}
-				/>
-			)}
-			{needsScrollingPlaceholder ? (
-				<SpotlightPlaceholder />
-			) : null}
-		</>
-	);
-};
-/* eslint-enable enact/prop-types */
-
-export default listItemsRenderer;
